feat: add Node.js improved FlameGraphs
Use the new StackProcessors introduced in `nflxprofile` to improve
visualization of Node.js FlameGraphs. This will add a new FlameGraph
setting called "Flavor", which is used to transform the FlameGraph
before showing it. The goal of transforming the FlameGraph is to improve
how we interpret it in some scenarios. The previous "Java Package Name"
option is merged into this new Flavor option.